Split up EventLinkNormalizer into sub files
